Updated HTML Files
As of 11/13/20 -- The home can now search independently by prof, class name, and crn. Case sensitivity has been addressed as well. All html files are now using an external CSS sheet which is also uploaded here. The buttons on the bottom are able still completely functional and each php page is styled now via an external style sheet. Make sure to have the updated php files in the same file otherwise it will not link. To use the external sheets just have them located in the same folder and they should be good to go.