Conflict notification functionality, updated home file
I created the functionality for finding conflicts, and partially integrated it into the project.  However, we still are only able to add one class and don't have a record of the classes that have been added, so I cannot fully integrate this functionality until that is available.  I cleaned up the home file, edited some functions, and made it so that when a class is added, it only displays at the proper time.  It is still not perfect, as classes that end on the half-hour still take up the whole hour, but this can be fixed by expanding the table of times to go in half-hour blocks, rather than by full hours.